+++ mfa/mixins.py
@@ -22,9 +22,10 @@ class MFAFormView(FormView):
         else:
             raise Http404
 
-    def pop_state(self):
+    @property
+    def challenge(self):
         try:
-            return self.request.session.pop('mfa_state')
+            return self.request.session['mfa_challenge']
         except KeyError as e:
             raise Http404 from e
 
@@ -35,10 +36,23 @@ class MFAFormView(FormView):
 
     def get_context_data(self, **kwargs):
         context = super().get_context_data(**kwargs)
-        context['mfa_data'], self.request.session['mfa_state'] = self.begin()
+        if 'mfa_data' not in context:
+            data, state = self.begin()
+            self.request.session['mfa_challenge'] = (data, state)
+            context['mfa_data'] = data
         return context
 
     def get_form(self):
         form = super().get_form()
         form.complete = self.complete
         return form
+
+    def form_invalid(self, form):
+        # do not generate a new challenge
+        return self.render_to_response(self.get_context_data(
+            form=form, mfa_data=self.challenge[0]
+        ))
+
+    def form_valid(self, form):
+        del self.request.session['mfa_challenge']
+        return super().form_valid(form)