+++ CHANGES.md
@@ -1,4 +1,4 @@
-0.2.0 (unreleased)
+0.2.0 (2021-07-02)
 ------------------
 
 -   Convert qrcode to template filter. In templates, change